+++ generics/arithmetic/arithmetic.go
@@ -3,10 +3,12 @@
 
 package arithmetic
 
-import "reflect"
+import (
+	"reflect"
+)
 
 // Operander is the interface that wraps the arithmetic representation method.
-// Is useful for adding custom behavior to named types when GetVal processes
+// It is useful for adding custom behavior to named types when GetVal processes
 // it, other wise, the underlying type is obtained and follows the extraction
 // rules.
 //
@@ -37,7 +39,7 @@ func Div(operanders ...interface{}) float64 {
 	return result
 }
 
-// Eq gets any number of elements and checks if are equals.
+// Eq gets any number of elements and checks if they are equals.
 func Eq(operanders ...interface{}) bool {
 	x := GetVal(operanders[0])
 
@@ -61,6 +63,23 @@ func Mul(operanders ...interface{}) float64 {
 	return result
 }
 
+// Ne gets any number of elements and checks if they are differents.
+func Ne(operanders ...interface{}) bool {
+	s := make(map[float64]struct{})
+
+	for _, v := range operanders {
+		ar := GetVal(v)
+
+		if _, ok := s[ar]; ok {
+			return false
+		}
+
+		s[ar] = struct{}{}
+	}
+
+	return true
+}
+
 // Sub gets any number of elements and returns their subtraction.
 func Sub(operanders ...interface{}) float64 {
 	result := GetVal(operanders[0])
@@ -72,8 +91,31 @@ func Sub(operanders ...interface{}) float64 {
 	return result
 }
 
-// GetVal extracts the arithmetic representation from any type. It is ruled by
-// the value extraction rules.
+/*
+GetVal extracts the arithmetic representation from any type. It is ruled by the
+value extraction rules.
+
+Value extraction rules
+
+1. Any element that satisfies the Operander interface will obtain its value
+from the Val method.
+
+2. Any element with a named type that doesn't satisfies the Operander interface
+will obtain its value from its underlying type.
+
+3. Boolean elements with a true value will be represented as 1, for false
+values they will be 0.
+
+4. Numeric elements (int, int8, int16, int32, int64, uint, uint8, uint16,
+uint32, uint64, float32, float64, complex64, complex128, byte, rune) will be
+converted to float64, but complex numbers will be represented by their real
+part in float64 form.
+
+5. Composed (arrays, maps, slices, strings, structs) and channel elements will
+be represented by their length (or their number of fields for structs).
+
+6. Any other element will be 0.
+*/
 func GetVal(operander interface{}) float64 {
 	if x, ok := operander.(Operander); ok {
 		return x.Val()