+++ generics/arithmetic/operander_example_test.go
@@ -9,8 +9,8 @@ import (
 	"github.com/ntrrg/ntgo/generics/arithmetic"
 )
 
-// BytesSum is a simple string implementation of Operander. Its arithmetic
-// representation is the sum of all its bytes.
+// BytesSum is a simple string implementation of arithmetic.Operander. Its
+// arithmetic representation is the sum of all its bytes values.
 type BytesSum string
 
 func (o BytesSum) Val() (r float64) {
@@ -24,8 +24,25 @@ func (o BytesSum) Val() (r float64) {
 func ExampleOperander() {
 	x := BytesSum("hello")
 	fmt.Println(x.Val())
-	fmt.Println(arithmetic.Add(x, 8))
+
+	fmt.Println(arithmetic.Add(x, BytesSum("world")))
+	fmt.Println(arithmetic.Sub(x, 32))
+	fmt.Println(arithmetic.Mul(x, "world"))
+	fmt.Println(arithmetic.Div(x, []byte{'M', 'A'}))
+
+	fmt.Println(arithmetic.Add(x))
+	fmt.Println(arithmetic.Sub(x))
+	fmt.Println(arithmetic.Mul(x))
+	fmt.Println(arithmetic.Div(x))
+
 	// Output:
 	// 532
-	// 540
+	// 1084
+	// 500
+	// 2660
+	// 266
+	// 532
+	// 532
+	// 532
+	// 532
 }